Refactor Login into login.rs
