+++ epp-client/src/lib.rs
@@ -108,6 +108,7 @@ pub mod domain;
 pub mod error;
 pub mod hello;
 pub mod host;
+pub mod login;
 pub mod message;
 pub mod request;
 pub mod response;
